refactor: move NextTalkNotifier to layout dir
