+++ src/pages/[eventAbbr]/ui/index.tsx
@@ -14,7 +14,7 @@ import { useSelector } from 'react-redux'
 import { pointEventSavingSelector } from '../../../store/appData'
 import { CircularProgress } from '@material-ui/core'
 import * as CommonStyled from '../../../styles/styled'
-import { NextTalkNotifier } from '../../../components/IvsPlayer/NextTalkNotifier'
+import { NextTalkNotifier } from '../../../components/Layout/NextTalkNotifier'
 
 const IndexPage: NextPage = () => {
   return withAuthProvider(<IndexMain />)
